added reset password test

+++ app/controllers/sessions_controller.rb
@@ -2,7 +2,7 @@ class SessionsController < ApplicationController
 
   def create
     @user = User.find_by(email:params[:session][:email])
-    if @user && @user.authenticate(params[:session][:password])
+    if @user && @user.activated? && @user.authenticate(params[:session][:password])
       login(@user)
       flash[:success] = "log in"
       redirect_to root_path
@@ -13,7 +13,7 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-    session.delete(:user_id)
+    session.delete(:user_id) if !session[:user_id].nil?
     redirect_to root_path
   end
 end
